+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `approvalsService.createApprovalAdditionalNote()`
 - `approvalsService.updateApprovalAdditionalNote()`
 - `approvalsService.deleteApprovalAdditionalNote()`
+- `authService.getCurrentFormsAppUser()`
 
 ## [1.2.2] - 2022-06-23